Update TameableDragonEntity.java
Hello,
Could you make it so we can change the amount of health and/or attack it has? I would like to add Dragon Mounts Legacy to a Modpack I've been creating. I don't release Modpacks as i make them for my videos and stuff. i would like them to last more than a day so I can have a companion for the 100 day videos.